requirements.txt
numpy==1.24.3
matplotlib==3.7.1
pandas==2.0.1
seaborn==0.12.2
pyyaml==6.0
scipy==1.10.1 